Added the terms of use page to the settings menu
