+++ app/src/main/java/com/example/chat_bot/Activities/Notification/Activity/Notification_UI/Notification_MainActivity.kt
@@ -9,11 +9,9 @@ import android.content.SharedPreferences
 import android.os.Build
 import android.os.Bundle
 import android.view.KeyEvent
-import android.view.WindowInsets
-import android.view.WindowManager
 import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
-import com.example.chat_bot.Activities.DashboardActivities.Settings
+import com.example.chat_bot.Activities.DashboardActivities.Settings.Settings
 import com.example.chat_bot.Activities.Notification.Activity.Notification_fragment.NotificationFragment
 import com.example.chat_bot.R
