+++ app/src/main/res/values-de/strings.xml
@@ -24,6 +24,59 @@
     <string name="flash_cards">Flash-Karten</string>
     <string name="no_flash_cards">Es werden keine Aktivitäten abgeschlossen, um die Karteikarten anzuzeigen. Versuchen Sie, einige Aktivitäten zu vervollständigen</string>
     <string name="erasmus_declaration">Von der Europäischen Union finanziert. Die geäußerten Ansichten und Meinungen entsprechen jedoch ausschließlich denen des Autors bzw. der Autoren und spiegeln nicht zwingend die der Europäischen Union oder der Europäischen Exekutivagentur für Bildung und Kultur (EACEA) wider. Weder die Europäische Union noch die EACEA können dafür verantwortlich gemacht werden.</string>
+    <string name="termsofuse">
+    Wir nehmen den Datenschutz ernst und verpflichten uns,
+    den Schutz und die Sicherheit Ihrer Daten zu gewährleisten.\n\n
+
+    Diese Datenschutzerklärung beschreibt, wie Ihre Daten erhoben,
+    verwendet und gespeichert werden, wenn Sie unsere mobile Anwendung
+    für das Forschungsprojekt nutzen.\n\n
+
+    1. Datenerhebung und -verwendung: Wir erheben minimale Daten, um ein Benutzerprofil zu erstellen,
+    einschließlich des von Ihnen gewählten Benutzernamens und Passworts
+    sowie Ihrer Geräte-ID. Diese Informationen werden ausschließlich zur
+    Erstellung Ihres Benutzerprofils innerhalb der Anwendung verwendet und
+    sind nicht mit personenbezogenen Daten verknüpft. Wir sammeln keine
+    persönlichen Daten wie Name, Adresse, Kontaktdaten oder andere sensible
+    Informationen.\n\n
+
+    2. Datenspeicherung und -verwaltung: Alle erfassten Daten werden von unserem Projektteam auf Servern in Deutschland
+    sicher gespeichert und verwaltet. Wir ergreifen strenge Maßnahmen, um Ihre Daten
+    vor unbefugtem Zugriff, Verwendung oder Offenlegung zu schützen.\n\n
+
+    3. Forschungszwecke: Die von der Anwendung gesammelten Nutzungsdaten werden ausschließlich für
+    Forschungszwecke im Zusammenhang mit dem Projekt verwendet. Diese Daten werden
+    analysiert, um Erkenntnisse über die Funktionalität der Anwendung, das Nutzerverhalten
+    und die Gesamtleistung zu gewinnen.\n\n
+
+    4. Aufbewahrung der Daten: Wir werden die gesammelten Daten für einen Zeitraum von einem Jahr nach Beendigung
+    des Projekts aufbewahren. Nach Ablauf dieses Zeitraums werden alle Daten sicher gelöscht,
+    so dass sie nicht mehr zugänglich oder nachvollziehbar sind.\n\n
+
+    5. Weitergabe von Daten: Wir geben keine Nutzerdaten an Dritte weiter, es sei denn, wir sind gesetzlich dazu
+    verpflichtet oder haben Ihre ausdrückliche Zustimmung. Nur unser Projektteam hat Zugriff
+    auf die Daten, und wir halten uns an strenge Vertraulichkeitsprotokolle.\n\n
+
+    6. Einverständnis: Durch die Nutzung unserer mobilen Anwendung erklären Sie sich mit der Erfassung und Nutzung
+    der in dieser Erklärung beschriebenen begrenzten Daten einverstanden. Ihre Teilnahme an dem
+    Forschungsprojekt ist vollkommen freiwillig, und Sie können Ihre Zustimmung jederzeit widerrufen,
+    indem Sie die Nutzung der Anwendung einstellen.\n\n
+
+    7. Datensicherheit: Wir setzen branchenübliche Sicherheitsmaßnahmen ein, um Ihre Daten vor unbefugtem Zugriff, Verlust,
+    Veränderung oder Offenlegung zu schützen. Bitte beachten Sie jedoch, dass keine Methode der Datenübertragung
+    über das Internet oder der elektronischen Speicherung völlig sicher ist, und obwohl wir uns bemühen, Ihre Daten
+    zu schützen, können wir deren absolute Sicherheit nicht garantieren.\n\n
+
+    Wenn Sie Bedenken oder Fragen zum Datenschutz oder zu dieser Erklärung haben, wenden Sie sich bitte an unser
+    Projektteam unter seeds.members@listserv.dfn.de.\n\n
+
+    Diese Datenschutzerklärung kann sich ändern, und alle Aktualisierungen werden den Nutzern über die mobile
+    Anwendung oder unsere Projektwebsite mitgeteilt. Wir empfehlen Ihnen, diese Erklärung regelmäßig zu lesen,
+    um sich über den Umgang mit Ihren Daten zu informieren.
+
+</string>
+
+
     <string name="lang_drop_down_label">Sprache</string>
     <string name="action_sign_in">Anmelden oder registrieren</string>
     <string name="action_sign_in_short">Anmelden</string>
@@ -113,4 +166,5 @@
     <string name="intro_page3">Die Benutzerstatistiken sind im Dashboard des Schülers verfügbar und zeigen den Gesamtprozentsatz der vom Schüler bereits abgeschlossenen Fächer an.</string>
     <string name="introduction_video">Einführungsvideo</string>
     <string name="intro_video">Das Einführungsvideo oben zeigt eine Demo der Funktionalität der App</string>
+    <string name="privacy_policy_statement_terms_of_use">Erklärung zum Datenschutz/Nutzungsbedingungen</string>
 </resources>
