+++ app/src/main/res/values-el/strings.xml
@@ -37,6 +37,9 @@
     <string name="flash_cards">Κάρτες Flash</string>
     <string name="no_flash_cards">Δεν ολοκληρώνονται δραστηριότητες για την προβολή των καρτών flash. Δοκιμάστε να ολοκληρώσετε κάποιες δραστηριότητες</string>
     <string name="erasmus_declaration">Με τη χρηματοδότηση της Ευρωπαϊκής Ένωσης. Οι απόψεις και οι γνώμες που διατυπώνονται εκφράζουν αποκλειστικά τις απόψεις των συντακτών και δεν αντιπροσωπεύουν κατ\'ανάγκη τις απόψεις της Ευρωπαϊκής Ένωσης ή του Ευρωπαϊκού Εκτελεστικού Οργανισμού Εκπαίδευσης και Πολιτισμού (EACEA). Η Ευρωπαϊκή Ένωση και ο EACEA δεν μπορούν να θεωρηθούν υπεύθυνοι για τις εκφραζόμενες απόψεις.</string>
+    <string name="termsofuse">"Λαμβάνουμε σοβαρά υπόψη το απόρρητο των δεδομένων και δεσμευόμαστε να διασφαλίσουμε την προστασία και την ασφάλεια των πληροφοριών σας.\n\n Η παρούσα δήλωση απορρήτου δεδομένων περιγράφει τον τρόπο με τον οποίο συλλέγονται, χρησιμοποιούνται και αποθηκεύονται τα δεδομένα σας όταν χρησιμοποιείτε την εφαρμογή μας για κινητά τηλέφωνα για το ερευνητικό έργο.\n\n  1. Συλλογή και χρήση δεδομένων: Συλλέγουμε ελάχιστα δεδομένα για τη δημιουργία ενός προφίλ χρήστη, συμπεριλαμβανομένων του επιλεγμένου ονόματος χρήστη και του κωδικού πρόσβασης, καθώς και του αναγνωριστικού της συσκευής σας. Οι πληροφορίες αυτές χρησιμοποιούνται αποκλειστικά για τη δημιουργία του προφίλ χρήστη σας εντός της εφαρμογής και δεν συνδέονται με οποιαδήποτε προσωπική πληροφορία. Δεν συλλέγουμε προσωπικά δεδομένα όπως όνομα, διεύθυνση, στοιχεία επικοινωνίας ή άλλες ευαίσθητες πληροφορίες.\n\n2. Αποθήκευση και διαχείριση δεδομένων: Όλα τα δεδομένα που συλλέγονται αποθηκεύονται και διαχειρίζονται με ασφάλεια από την ομάδα έργου μας σε διακομιστές που βρίσκονται στη Γερμανία. Διατηρούμε αυστηρά μέτρα για την προστασία των δεδομένων σας από μη εξουσιοδοτημένη πρόσβαση, χρήση ή αποκάλυψη.\n\n3. Ερευνητικοί σκοποί: Τα δεδομένα χρήσης που συλλέγονται από την εφαρμογή χρησιμοποιούνται αποκλειστικά για ερευνητικούς σκοπούς που σχετίζονται με το έργο. Τα δεδομένα αυτά θα αναλυθούν για να αποκτηθούν γνώσεις σχετικά με τη λειτουργικότητα της εφαρμογής, τη συμπεριφορά των χρηστών και τη συνολική απόδοση.\n\n4. Διατήρηση δεδομένων: Θα διατηρήσουμε τα δεδομένα που συλλέγονται για περίοδο ενός έτους μετά την ολοκλήρωση του έργου. Μετά το πέρας αυτής της περιόδου, όλα τα δεδομένα θα διαγραφούν με ασφάλεια, διασφαλίζοντας ότι δεν είναι πλέον προσβάσιμα ή ανιχνεύσιμα.\n\n5. Κοινή χρήση δεδομένων: Δεν μοιραζόμαστε δεδομένα χρηστών με τρίτους, εκτός εάν απαιτείται από το νόμο ή με τη ρητή συγκατάθεσή σας. Η ομάδα έργου μας είναι η μόνη οντότητα που έχει πρόσβαση στα δεδομένα και τηρούμε αυστηρά πρωτόκολλα εμπιστευτικότητας.\n\n6. Συναίνεση: Χρησιμοποιώντας την εφαρμογή μας για κινητά, συναινείτε στη συλλογή και χρήση των περιορισμένων δεδομένων που περιγράφονται στην παρούσα δήλωση. Η συμμετοχή σας στο ερευνητικό έργο είναι απολύτως εθελοντική και μπορείτε να ανακαλέσετε τη συγκατάθεσή σας ανά πάσα στιγμή διακόπτοντας τη χρήση της εφαρμογής.\n\n7. Ασφάλεια δεδομένων: Εφαρμόζουμε μέτρα ασφαλείας σύμφωνα με τα πρότυπα του κλάδου για την προστασία των δεδομένων σας από μη εξουσιοδοτημένη πρόσβαση, απώλεια, αλλοίωση ή αποκάλυψη. Ωστόσο, πρέπει να γνωρίζετε ότι καμία μέθοδος μετάδοσης δεδομένων μέσω του διαδικτύου ή ηλεκτρονικής αποθήκευσης δεν είναι απολύτως ασφαλής, και παρόλο που προσπαθούμε να προστατεύσουμε τα δεδομένα σας, δεν μπορούμε να εγγυηθούμε την απόλυτη ασφάλειά τους.\n\nΓια οποιεσδήποτε ανησυχίες ή ερωτήσεις σχετικά με το απόρρητο των δεδομένων ή την παρούσα δήλωση, παρακαλούμε επικοινωνήστε με την ομάδα έργου μας στη διεύθυνση seeds.members@listserv.dfn.de.\n
+
+Η παρούσα δήλωση απορρήτου δεδομένων υπόκειται σε αλλαγές και τυχόν ενημερώσεις θα γνωστοποιούνται στους χρήστες μέσω της εφαρμογής για κινητά ή του ιστότοπου του έργου μας. Συνιστούμε να ελέγχετε αυτή τη δήλωση περιοδικά για να ενημερώνεστε σχετικά με τον τρόπο χειρισμού των δεδομένων σας. </string>
     <string name="lang_drop_down_label">γλώσσες</string>
     <string name="lang_drop_down_text">Επιλογή γλώσσας</string>
     <string name="register_account">Δημιούργησε έναν λογαριασμό</string>
@@ -120,6 +123,7 @@
     <string name="intro_page3">Τα στατιστικά στοιχεία χρήστη είναι διαθέσιμα στο ταμπλό του μαθητή, τα οποία εμφανίζουν το συνολικό ποσοστό των θεμάτων που έχει ήδη ολοκληρώσει ο μαθητής.</string>
     <string name="introduction_video">Εισαγωγικό βίντεο</string>
     <string name="intro_video">Το παραπάνω εισαγωγικό βίντεο παρουσιάζει μια επίδειξη της λειτουργικότητας της εφαρμογής</string>
+    <string name="privacy_policy_statement_terms_of_use">Δήλωση πολιτικής απορρήτου / Όροι χρήσης</string>
 
 
 </resources>
