+++ app/src/main/res/values-es/strings.xml
@@ -25,6 +25,20 @@
     <string name="flash_cards">Fichas</string>
     <string name="no_flash_cards">No se completa ninguna actividad para mostrar las tarjetas flash. Intenta completar algunas actividades</string>
     <string name="erasmus_declaration">Financiado por la Unión Europea. Las opiniones y puntos de vista expresados solo comprometen a su(s) autor(es) y no reflejan necesariamente los de la Unión Europea o los de la Agencia Ejecutiva Europea de Educación y Cultura (EACEA). Ni la Unión Europea ni la EACEA pueden ser considerados responsables de ellos.</string>
+    <string name="termsofuse">
+    "Nos tomamos muy en serio la privacidad de los datos y nos comprometemos a garantizar la protección y seguridad de su información.\n\n
+    Esta declaración de privacidad de datos describe cómo se recopilan, utilizan y almacenan sus datos cuando utiliza nuestra aplicación móvil para el proyecto de investigación.\n
+    1. Recogida y uso de datos: Recopilamos datos mínimos para crear un perfil de usuario, incluyendo su nombre de usuario y contraseña seleccionados, así como el ID de su dispositivo. Esta información se utiliza únicamente para establecer su perfil de usuario dentro de la aplicación y no está vinculada a ninguna información de identificación personal. No recopilamos ningún dato personal como nombre, dirección, datos de contacto o cualquier otra información sensible.\n
+    2. Almacenamiento y gestión de datos: Todos los datos recopilados son almacenados y gestionados de forma segura por nuestro equipo de proyecto en servidores ubicados en Alemania. Mantenemos medidas estrictas para salvaguardar sus datos del acceso, uso o divulgación no autorizados.\n\n
+    3. Fines de investigación: Los datos de uso recogidos de la aplicación se utilizan únicamente con fines de investigación relacionados con el proyecto. Estos datos se analizarán para obtener información sobre la funcionalidad de la aplicación, el comportamiento de los usuarios y el rendimiento general.\n
+    4. Retención de datos: Conservaremos los datos recopilados durante un año tras la finalización del proyecto. Tras este periodo, todos los datos se eliminarán de forma segura, garantizando que ya no sean accesibles ni rastreables.\n
+    5. Intercambio de datos: No compartimos ningún dato de los usuarios con terceros, excepto cuando lo exige la ley o con su consentimiento explícito. Nuestro equipo de proyecto es la única entidad con acceso a los datos, y nos adherimos a estrictos protocolos de confidencialidad.\n\n
+    6. Consentimiento: Al utilizar nuestra aplicación móvil, usted consiente la recopilación y el uso de los datos limitados descritos en esta declaración. Su participación en el proyecto de investigación es totalmente voluntaria, y puede retirar su consentimiento en cualquier momento dejando de utilizar la aplicación.\n
+    7. Seguridad de los datos: Empleamos medidas de seguridad estándar de la industria para proteger sus datos contra el acceso no autorizado, pérdida, alteración o divulgación. Sin embargo, tenga en cuenta que ningún método de transmisión de datos por Internet o de almacenamiento electrónico es totalmente seguro y, aunque nos esforzamos por proteger sus datos, no podemos garantizar su seguridad absoluta.\n\n
+    Para cualquier duda o consulta relacionada con la privacidad de los datos o esta declaración, póngase en contacto con nuestro equipo de proyectos en seeds.members@listserv.dfn.de.\n
+    Esta declaración de privacidad de datos está sujeta a cambios, y cualquier actualización se comunicará a los usuarios a través de la aplicación móvil o del sitio web de nuestro proyecto. Le recomendamos que consulte esta declaración periódicamente para mantenerse informado sobre el tratamiento de sus datos."
+</string>
+
     <string name="lang_drop_down_label">Idiomas</string>
     <string name="Username">Usuario</string>
     <string name="register_account">Crea una cuenta</string>
@@ -116,4 +130,5 @@
     <string name="intro_page3">Las estadísticas del usuario están disponibles en el panel de control del estudiante, que muestra el porcentaje total de asignaturas completadas por el estudiante.</string>
     <string name="introduction_video">Vídeo de introducción</string>
     <string name="intro_video">El vídeo de introducción anterior muestra una demostración de la funcionalidad de la aplicación</string>
+    <string name="privacy_policy_statement_terms_of_use">Política de privacidad / Condiciones de uso</string>
 </resources>
